\name{countComm-package}
\alias{countComm-package}
\alias{countComm}
\docType{package}
\title{
\packageTitle{countComm}
}
\description{
\packageDescription{countComm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{countComm}
\packageIndices{countComm}
}
\author{
\packageAuthor{countComm}

Maintainer: \packageMaintainer{countComm}
}
\references{
Blanchet F. G., P. Legendre, and F. He. 2016. A new cost-effective approach to survey ecological communities. \emph{Oikos} 125: 75-987,. doi: 10.1111/oik.02838.
}
\keyword{ package }
